+++ src/festim/subdomain/__init__.py
@@ -1,7 +1,6 @@
 from .interface import Interface
 from .surface_subdomain import SurfaceSubdomain, SurfaceSubdomain1D
-from .volume_subdomain import VolumeSubdomain
-from .volume_subdomain_1d import VolumeSubdomain1D
+from .volume_subdomain import VolumeSubdomain, VolumeSubdomain1D
 
 __all__ = [
     "Interface",
