+++ src/festim/subdomain/volume_subdomain.py
@@ -1,6 +1,7 @@
 import dolfinx
 from dolfinx.mesh import Mesh, locate_entities
 import numpy as np
+from numpy import typing as npt
 from typing import Callable
 
 from festim.helpers_discontinuity import transfer_meshtags_to_submesh
@@ -69,7 +70,7 @@ class VolumeSubdomain:
             mesh, tag, self.submesh, self.v_map, self.submesh_to_mesh
         )
 
-    def locate_subdomain_entities(self, mesh: Mesh) -> np.NDArray[np.int32]:
+    def locate_subdomain_entities(self, mesh: Mesh) -> npt.NDArray[np.int32]:
         """Locates all cells in subdomain borders within domain
 
         Args:
@@ -83,6 +84,41 @@ class VolumeSubdomain:
 
         entities = locate_entities(mesh, mesh.topology.dim, self.locator)
         return entities
+
+
+class VolumeSubdomain1D(VolumeSubdomain):
+    """
+    Volume subdomain class for 1D cases
+
+    Args:
+        id (int): the id of the volume subdomain
+        borders (list of float): the borders of the volume subdomain
+        material (festim.Material): the material of the volume subdomain
+
+    Attributes:
+        id (int): the id of the volume subdomain
+        borders (list of float): the borders of the volume subdomain
+        material (festim.Material): the material of the volume subdomain
+
+    Examples:
+
+        .. testsetup:: VolumeSubdomain1D
+
+            from festim import VolumeSubdomain1D, Material
+            my_mat = Material(D_0=1, E_D=1, name="test_mat")
+
+        .. testcode:: VolumeSubdomain1D
+
+            VolumeSubdomain1D(id=1, borders=[0, 1], material=my_mat)
+    """
+
+    def __init__(self, id, borders, material) -> None:
+        super().__init__(
+            id,
+            material,
+            locator=lambda x: np.logical_and(x[0] >= borders[0], x[0] <= borders[1]),
+        )
+        self.borders = borders
 
 
 def find_volume_from_id(id: int, volumes: list):
